Warnings

+++ trunk/Source/Utils/AmjuRand.cpp
@@ -1,5 +1,6 @@
 #include <AmjuFirst.h>
 #include <stdlib.h>
+#include <time.h>
 #include "AmjuRand.h"
 #include <AmjuFinal.h>
 
@@ -8,7 +9,7 @@ namespace Amju
 void Randomise()
 {
   time_t tt;
-  srand(time(&tt));
+  srand((unsigned int)time(&tt));
 }
 
 float Rnd(float min, float max)
